+++ src/types/locale.d.ts
@@ -11,30 +11,18 @@ export type TranslationKeys = {
   'network.deleteSuccess': 'emoji' | 'user' | 'deleted' | 'total';
   'network.editSuccess': 'emoji' | 'user' | 'edited' | 'total';
   'network.welcome': 'emoji' | 'user' | 'hub' | 'totalServers' | 'channel' | 'rules_command';
-  'network.nsfw.title': never;
-  'network.nsfw.description': 'rules_command';
-  'network.nsfw.footer': never;
   'network.onboarding.embed.title': 'hubName';
   'network.onboarding.embed.description': 'docs_link';
   'network.onboarding.embed.footer': 'version';
   'network.onboarding.inProgress': 'emoji' | 'channel';
-  'blacklist.embed.user': 'emoji';
-  'blacklist.embed.userValue': 'user';
-  'blacklist.embed.userAlreadyBlacklisted': 'user';
-  'blacklist.embed.server': 'emoji';
-  'blacklist.embed.serverValue': 'server';
-  'blacklist.embed.serverAlreadyBlacklisted': 'server';
-  'blacklist.embed.footer': never;
+  'blacklist.success': 'emoji' | 'name';
+  'blacklist.removed': 'emoji' | 'name';
   'blacklist.modal.reason.label': never;
   'blacklist.modal.reason.placeholder': never;
   'blacklist.modal.duration.label': never;
   'blacklist.modal.duration.placeholder': never;
-  'blacklist.user.success': 'emoji' | 'username';
-  'blacklist.user.removed': 'emoji' | 'username';
   'blacklist.user.alreadyBlacklisted': 'emoji';
   'blacklist.user.easterEggs.blacklistBot': never;
-  'blacklist.server.success': 'emoji' | 'server';
-  'blacklist.server.removed': 'emoji' | 'server';
   'blacklist.server.alreadyBlacklisted': 'emoji';
   'blacklist.server.unknownError': 'server';
   'blacklist.list.user': 'id' | 'moderator' | 'reason' | 'expires';
